Port HQ minimap texture generation to mod system
