+++ mods/hq_minimap/mod.json
@@ -1,0 +1,7 @@
+{
+  "id": "dev.twilitrealm.hq_minimap",
+  "name": "HQ Minimap",
+  "version": "1.0.0",
+  "author": "Twilit Realm",
+  "description": "Higher-quality textures for minimap rendering."
+}
